% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{assign_sample_numbers}
\alias{assign_sample_numbers}
\title{Assign sample numbers.}
\usage{
assign_sample_numbers(raw_data)
}
\arguments{
\item{raw_data}{A \code{data.frame} of data returned by \code{\link{process_directory}}.}
}
\value{
The data with new \code{Sample_number} and \code{Elapsed_seconds} columns.
}
\description{
Assign sample numbers.
}
