% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_data}
\alias{clean_data}
\title{Clean the data: create \code{DATETIME} field, remove fractional and unwanted valves}
\usage{
clean_data(raw_data, tz = "", remove_valves = c())
}
\arguments{
\item{raw_data}{A \code{data.frame} of data returned by \code{\link{process_directory}}.}

\item{tz}{Timezone of Picarro timestamps, e.g. "America/New_York", character.
See \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}.}

\item{remove_valves}{An optional vector of integer valve numbers to remove
from the data, e.g. ambient ports.}
}
\value{
The cleaned data.
}
\description{
Clean the data: create \code{DATETIME} field, remove fractional and unwanted valves
}
