% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{compute_flux}
\alias{compute_flux}
\title{Compute respiration flux from concentration data.}
\usage{
compute_flux(time, gas_ppm, volume_cm3, tair_C, pressure_kPa = 101.325,
  debug_plot = FALSE)
}
\arguments{
\item{time}{Time (typically in seconds); numeric vector}

\item{gas_ppm}{Gas (e.g. CO2 or CH4) concentrations, ppmv; numeric vector}

\item{volume_cm3}{System volume, cm3; numeric}

\item{tair_C}{Air temperature, C; numeric}

\item{pressure_kPa}{Air pressure, kPa; numeric}

\item{debug_plot}{Produce a plot of data and fitted model? Logical}
}
\value{
The gas flux, µmol per unit time.
}
\description{
Compute respiration flux from concentration data.
}
\details{
We derive the rate of concentration change via the slope of the
linear regression of all data points of concentrations versus time.
One problem with this method is that the linearity attributed to the
regression is incorrect in a strict sense, since the gradients in
water vapor and CO2 decrease over time in a closed chamber (see
discussion in Steduto et al. below). This method should thus not be
used for long measurement periods.
}
\note{
The computed is \emph{not} area- or mass-corrected.
}
\examples{
# ten-second CO2 rise at room temperature
compute_flux(1:10, 401:410, volume_cm3 = 2000, tair_C = 20)
}
\references{
\itemize{
\item{Steduto et al. (2002), "Automated closed-system canopy-chamber for continuous
field-crop monitoring of CO2 and H2O fluxes" AFM 111:171-186
 \url{http://dx.doi.org/10.1016/S0168-1923(02)00023-0}.}
\item{Campbell and Normal (1998), An Introduction to Environmental Biophysics,
 ISBN 978-1-4612-1626-1. New York: Springer-Verlag, 286 pp.}
 \item{LI-COR Biosciences (2015), Using the LI-8100A Soil Gas Flux System & the LI-8150
 Multiplexer, 984-11123. 238 pp.}
}
}
