% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluxes.R
\name{cumulative_evolution}
\alias{cumulative_evolution}
\title{Compute cumulative C evolved}
\usage{
cumulative_evolution(time, flux, interpolate = TRUE)
}
\arguments{
\item{time}{Time, typically s or hr but arbitrary; numeric}

\item{flux}{Flux, in units of X/unit time; numeric}

\item{interpolate}{Interpolate missing values? Logical}
}
\value{
A vector of cumulative values of X evolved.
}
\description{
Compute cumulative C evolved
}
\details{
X evolved at timestep \emph{i} is the
lag between \code{time[i]} and \code{time[i-1]} multiplied by the mean of
\code{flux[i]} and \code{flux[i-1]}.
The first timestamp value is considered as time zero, so its corresponding
cumulative value will always be zero.
}
\examples{
cumulative_evolution(1:4, c(1, 1, NA, 1))
cumulative_evolution(1:4, c(1, 1, 2, 1))
}
