% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{process_directory}
\alias{process_directory}
\title{Scan a directory and read all Picarro files in it.}
\usage{
process_directory(path, tz = "", recursive = TRUE)
}
\arguments{
\item{path}{Folder to process, character.}

\item{tz}{Timezone to use for timestamps, e.g. "America/New_York", character.
See \url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones}.}

\item{recursive}{Scan (via \code{\link{list.files}}) subfolders? Logical.}
}
\value{
A \code{data.frame} of the data.
}
\description{
Scan a directory and read all Picarro files in it.
}
\note{
This can read plain-text or compressed (via \code{zip} or \code{gz}) files.
}
