% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_picarro_file}
\alias{read_picarro_file}
\title{Read a single Picarro output file.}
\usage{
read_picarro_file(filename)
}
\arguments{
\item{filename}{Fully qualified filename to read.}
}
\value{
A \code{data.frame} of the data.
}
\description{
Read a single Picarro output file.
}
\note{
This can read plain-text or compressed (via \code{zip} or \code{gz}) files.
}
